+++ src/Player.cpp
@@ -417,16 +417,16 @@ vector<Territory*>& Player::toAttack()
     vector<Territory*> toAttack = vector<Territory*>();
 
     // loop through each owned territory
-    for (int i = 0; i < this->getPlayerTerritories().size(); i++){
+    for (size_t i = 0; i < this->getPlayerTerritories().size(); i++){
         // get the adjacent territories to the current owned territory
         vector<Territory*> adjacentList = this->getPlayerTerritories().at(i)->neighbors;
         // loop through each adjacent territories
-        for (int j = 0; j < adjacentList.size(); j++) {
+        for (size_t j = 0; j < adjacentList.size(); j++) {
             bool found = false;
             //Only add to list if owned by enemy
             if (adjacentList.at(j)->player != this) {
                 // for each territory that has been added to toAttack
-                for (int k = 0; k < toAttack.size(); k++) {
+                for (size_t k = 0; k < toAttack.size(); k++) {
                     // make sure the current adjacentTerritory is not already in the list
                     if (toAttack.at(k)->name.compare(adjacentList.at(j)->name) == 0) {
                         found = true;
